requested changes

+++ apps/server/src/mindmap/mindmap.service.ts
@@ -32,21 +32,37 @@ export class MindmapService {
         path: doc.path,
         name: doc._id,
       });
-      await this.repository.findByIdAndUpdate(doc._id, {path: `${doc.path}/${doc._id}`})
+      await this.repository.findByIdAndUpdate(doc._id, {
+        path: `${doc.path}/${doc._id}`,
+      });
       return doc;
     } catch (error) {
-      this.logger.error('Error creating Mindmap', error);
+      this.logger.error(
+        'Error creating Mindmap',
+        `mindmapService.create`,
+        error,
+      );
       throw new Error();
     }
   }
 
-  findAll() {
-    return `This action returns all mindmap`;
+  async findAll(query) {
+    try {
+      return await this.repository.find(query);
+    } catch (error) {
+      this.logger.error(
+        'Error finding all mindmap of one workspace',
+        `mindmapService.findAll`,
+        error,
+      );
+      throw new Error();
+    }
   }
 
-  async findOne(id: string): Promise<any> {
+  async findOne(id: Types.ObjectId): Promise<any> {
     try {
-      const key = `temp/${id}`;
+      const doc = await this.repository.findById(id);
+      const key = doc.path;
       const response = await this.fileService.getFile(key);
 
       // Convert the response body stream to a string
@@ -64,17 +80,27 @@ export class MindmapService {
       const data = await streamToString(response.Body as Readable);
       return JSON.parse(data);
     } catch (error) {
-      this.logger.error('Error fetching mindmap from bucket', error);
+      this.logger.error(
+        'Error fetching mindmap from bucket',
+        `mindmapService.findOne`,
+        error,
+      );
       throw new Error();
     }
   }
 
-  async update(id: Types.ObjectId, updateMindmapDto: UpdateMindmapDto): Promise<Mindmap> {
+  async update(
+    id: Types.ObjectId,
+    updateMindmapDto: UpdateMindmapDto,
+  ): Promise<Mindmap> {
     try {
       return await this.repository.findByIdAndUpdate(id, updateMindmapDto);
-    }
-    catch(error) {
-      this.logger.error('Error updating mindmap from db', error);
+    } catch (error) {
+      this.logger.error(
+        'Error updating mindmap from db',
+        `mindmapService.update`,
+        error,
+      );
       throw new Error();
     }
   }
@@ -84,10 +110,15 @@ export class MindmapService {
       const doc = await this.repository.findById(id);
       const key = doc.path;
       const res = await this.fileService.delete(key);
-      await this.repository.findByIdAndDelete(id);
-      return;
+      if (res) {
+        await this.repository.findByIdAndDelete(id);
+      }
     } catch (error) {
-      this.logger.error('Error deleting mindmap', error);
+      this.logger.error(
+        'Error deleting mindmap',
+        `mindmapService.remove`,
+        error,
+      );
       throw new Error();
     }
   }
@@ -101,12 +132,16 @@ export class MindmapService {
       const mindmapData = {
         name: doc._id,
         content: saveMindmap.content,
-        path: 'temp',
+        path: doc.path,
       };
       await this.upload(mindmapData);
       return;
     } catch (error) {
-      this.logger.error('Error updating mindmap in bucket', error);
+      this.logger.error(
+        'Error updating mindmap in bucket',
+        `mindmapService.saveMindmap`,
+        error,
+      );
       throw new Error();
     }
   }
@@ -120,7 +155,11 @@ export class MindmapService {
       };
       return await this.fileService.uploadJSON(mindmapPayload, mindmap.path);
     } catch (error) {
-      this.logger.error('Error uploading mindmap to bucket', error);
+      this.logger.error(
+        'Error uploading mindmap to bucket',
+        `mindmapService.upload`,
+        error,
+      );
       throw new Error();
     }
   }
